Crate flowistry_pdg_construction

Compute program dependence graphs (PDG) for a function call graph.

Re-exports

pub use callback::CallChangeCallback;

pub use callback::CallChangeCallbackFn;

pub use callback::CallChanges;

pub use callback::CallInfo;

pub use callback::InlineMissReason;

pub use callback::SkipCall;

Modules

analysis 🔒

callback

Callbacks to influence graph construction and their supporting types.

constants

source_access

utils

Macros

debug_assert_resolved

Structs

DepEdge

An edge in the program dependence graph.

DepNode

A node in the program dependency graph.

MemoPdgConstructor

A memoizing constructor of PDGs.

OneHopLocation

Usually a location in a MIR body but can also cross “one hop” into a called function.

PartialGraph

VisitDriver

Enums

CallingConvention

Describes how the formal parameters of a given function call relate to the actual parameters.

DepEdgeKind

A kind of edge in the program dependence graph.

DepNodeKind

Use

Traits

Visitor

Functions

determine_async

Try to interpret this function as an async function.

is_async_trait_fn

Does this function have a structure as created by the #[async_trait] macro

match_async_trait_assign

Type Aliases

Node