flowistry_pdg_construction::analysis
Module global

Re-exports

pub use graph_elems::DepEdge;

pub use graph_elems::DepEdgeKind;

pub use graph_elems::DepNode;

pub use graph_elems::DepNodeKind;

pub use graph_elems::OneHopLocation;

pub use graph_elems::Use;

pub use partial_graph::Node;

pub use partial_graph::NodeKey;

pub use partial_graph::PartialGraph;

Modules

call_tree_visit

Machinery for visiting a monomorphized tree with access to intermediate PDGs.

graph_elems 🔒

The representation of the PDG.

partial_graph 🔒

Structs

GraphSizeEstimator 🔒

HumanInt 🔒

MemoPdgConstructor

A memoizing constructor of PDGs.

Enums

ConstructionResult

Input 🔒

Type Aliases

LocalAnalysisResults 🔒

PdgCache

Stores PDG’s we have already computed and which we know we can use again given a certain key.

PdgCacheKey

How we are indexing into PdgCache