flowistry_pdg_construction
Module async_support

Structs

AsyncFnPollEnv

Context for a call to Future::poll, when called on a future created via an async fn or an async block.

AsyncInfo 🔒

Stores ids that are needed to construct projections around async functions.

Enums

AsyncType

Describe in which way a function is async.

Functions

determine_async

Try to interpret this function as an async function.

get_async_generator 🔒

has_async_trait_signature 🔒

is_async_fn_or_block

Does this instance refer to an async fn or async {}.

is_async_trait_fn

Does this function have a structure as created by the #[async_trait] macro

match_async_trait_assign

match_pin_box_dyn_ty 🔒

try_as_async_trait_function