Module flowistry_pdg_construction::body_cache

Structs

BodyCache

Allows loading bodies from previosly written artifacts.

CachedBody

A mir Body and all the additional borrow checking facts that our points-to analysis needs.

DumpingVisitor 🔒

A visitor to collect all bodies in the crate and write them to disk.

FlowistryFacts

The subset of borrowcheck facts that the points-to analysis (flowistry) needs.

Constants

INTERMEDIATE_ARTIFACT_EXT 🔒

Functions

clean_undecodable_data_from_body 🔒

Some data in a Body is not cross-crate compatible. Usually because it involves storing a LocalDefId. This function makes sure to sanitize those out.

dump_mir_and_borrowck_facts

A complete visit over the local crate items, collecting all bodies and calculating the necessary borrowcheck facts to store for later points-to analysis.

intermediate_out_dir

Create the name of the file in which to store intermediate artifacts.

load_body_and_facts 🔒

Try to load a CachedBody for this id.

local_or_remote_paths

Get the path where artifacts from this crate would be stored. Unlike TyCtxt::crate_extern_paths this function does not crash when supplied with LOCAL_CRATE.

Type Aliases

LocationIndex