flowistry_pdg_construction::body_cache
Function local_or_remote_paths

pub fn local_or_remote_paths(
    krate: CrateNum,
    tcx: TyCtxt<'_>,
    ext: &str,
) -> Vec<PathBuf>

Get the path where artifacts from this crate would be stored. Unlike TyCtxt::crate_extern_paths this function does not crash when supplied with LOCAL_CRATE.