flowistry_pdg_construction
Module call_tree_visit

Machinery for visiting a monomorphized tree with access to intermediate PDGs.

Each traversal is comprised of two components:

A VisitDriver that drives the traversal and manages associated state. Mainly the currently visited function and the call stack.

An object implementing Visitor which contains the custom logic of the traversal.

The relation of the to is such that for each visitation step the *_visit method on Visitor is called first. It should then call the corresponding *_visit method on the VisitDriver which will take care of the recursion.

Structs

VisitDriver

Traits

Visitor