flowistry_pdg_construction
Module construct

Structs

GraphAssembler 🔒

GraphSizeEstimator 🔒

HumanInt 🔒

MemoPdgConstructor

A memoizing constructor of PDGs.

Enums

ConstructionResult

Inputs 🔒

Functions

globalize_edge 🔒

globalize_location 🔒

globalize_node 🔒

Type Aliases

LocalAnalysisResults 🔒

PdgCache

Stores PDG’s we have already computed and which we know we can use again given a certain key.

PdgCacheKey

How we are indexing into PdgCache