flowistry_pdg_construction::construct
Struct GraphAssembler

struct GraphAssembler<'tcx> {
    graph: DiGraph<DepNode<'tcx, CallString>, DepEdge<CallString>>,
    nodes: FxHashMap<DepNode<'tcx, CallString>, NodeIndex>,
    control_inputs: Box<[(NodeIndex, DepEdge<CallString>)]>,
}

Fields
graph: DiGraph<DepNode<'tcx, CallString>, DepEdge<CallString>>nodes: FxHashMap<DepNode<'tcx, CallString>, NodeIndex>control_inputs: Box<[(NodeIndex, DepEdge<CallString>)]>
Implementations

impl<'tcx> GraphAssembler<'tcx>

fn new() -> Self

fn add_node(&mut self, node: DepNode<'tcx, CallString>) -> NodeIndex

fn with_new_ctr_inputs<'c, F, R, K>( &mut self, vis: &mut VisitDriver<'tcx, 'c, K>, new_ctrl_inputs: &[(DepNode<'tcx, OneHopLocation>, DepEdge<OneHopLocation>)], f: F, ) -> R
where F: FnOnce(&mut Self, &mut VisitDriver<'tcx, 'c, K>) -> R,

Forwarding of control flow. It is sound to replace the control inputs here rather than extend them because we are guaranteed that these new nodes are connected to the old ctrl nodes, possibly transitively.

Each node in our graph is either connected to a local control flow node or to the ones coming from the parent, which is established by the visit_partial_graph function. By induction all nodes, including these control flow sources are connected to the old ctrl inputs.