flowistry_pdg_construction::construct
Type Alias PdgCache

pub type PdgCache<'tcx, K> = Rc<TwoLevelCache<Instance<'tcx>, K, Option<PartialGraph<'tcx, K>>>>;

Stores PDG’s we have already computed and which we know we can use again given a certain key.

Aliased Type

struct PdgCache<'tcx, K> { /* private fields */ }