flowistry_pdg_construction
Module encoder

Readers and writers for the intermediate artifacts we store per crate.

Most of this code is adapted/copied from EncodeContext and DecodeContext in rustc_metadata.

We use a lot of min_specialization here to change how DefIds, Spans and such are encoded since their default implementations are panics.

Specifically for CrateNum (e.g. DefId also), we use stable crate hashes. These appear to work fine, however I am not sure they are guaranteed to be stable across different crates. Rustc itself uses an explicit remapping replying on CrateMetadataRef, which we can construct but not use (relevant functions are hidden).

Note that we encode AllocIds simply as themselves. This is possibly incorrect but we’re not really relying on this information at the moment so we are not investing in it.

Macros

decoder_methods 🔒

encoder_methods 🔒

Structs

ParalegalDecoder

Something that implements TyDecoder.

ParalegalEncoder

A structure that implements TyEncoder for us.

Constants

CLEAR_CROSS_CRATE 🔒

Whatever can’t survive the crossing we need to live without.

TAG_ENCODE_LOCAL 🔒

TAG_ENCODE_REMOTE 🔒

TAG_PARTIAL_SPAN 🔒

TAG_VALID_SPAN_FULL 🔒

Functions

decode_from_file

Convenience function that decodes a value from a file.

encode_to_file

Convenience function that encodes some value to a file.

path_in_real_path 🔒

Which path in a RealFileName do we care about?