flowistry_pdg_construction::graph
Struct PartialGraph

pub struct PartialGraph<'tcx, K> {
    pub(crate) nodes: FxHashSet<DepNode<'tcx, OneHopLocation>>,
    pub(crate) edges: FxHashSet<(DepNode<'tcx, OneHopLocation>, DepNode<'tcx, OneHopLocation>, DepEdge<OneHopLocation>)>,
    pub(crate) generics: GenericArgsRef<'tcx>,
    pub(crate) def_id: DefId,
    arg_count: usize,
    local_decls: IndexVec<Local, LocalDecl<'tcx>>,
    pub(crate) k: K,
    pub(crate) inlined_calls: Vec<(Location, Instance<'tcx>, K, Vec<(DepNode<'tcx, OneHopLocation>, DepEdge<OneHopLocation>)>)>,
}

Fields
nodes: FxHashSet<DepNode<'tcx, OneHopLocation>>edges: FxHashSet<(DepNode<'tcx, OneHopLocation>, DepNode<'tcx, OneHopLocation>, DepEdge<OneHopLocation>)>generics: GenericArgsRef<'tcx>def_id: DefIdarg_count: usizelocal_decls: IndexVec<Local, LocalDecl<'tcx>>k: Kinlined_calls: Vec<(Location, Instance<'tcx>, K, Vec<(DepNode<'tcx, OneHopLocation>, DepEdge<OneHopLocation>)>)>
Implementations

impl<'tcx, K: Hash + Eq + Clone> PartialGraph<'tcx, K>

fn modular_mutation_visitor<'a, 'mir>( &'a mut self, results: &'a Results<'tcx, &'mir LocalAnalysis<'tcx, 'mir, K>>, state: &'a InstructionState<'tcx>, ) -> ModularMutationVisitor<'a, 'tcx, impl FnMut(Location, Mutation<'tcx>) + use<'a, 'tcx, 'mir, K>>

fn handle_as_inline<'a>( &mut self, results: &Results<'tcx, &'a LocalAnalysis<'tcx, 'a, K>>, state: &InstructionState<'tcx>, terminator: &Terminator<'tcx>, location: Location, ) -> bool

returns whether we were able to successfully handle this as inline

fn register_mutation( &mut self, results: &Results<'tcx, &'_ LocalAnalysis<'tcx, '_, K>>, state: &InstructionState<'tcx>, inputs: Inputs<'tcx>, mutated: Either<Place<'tcx>, DepNode<'tcx, OneHopLocation>>, location: Location, target_use: TargetUse, )
where K: Hash + Eq + Clone,

impl<'tcx, K: Clone + Hash + Eq> PartialGraph<'tcx, K>

pub fn to_petgraph<'c>( &self, memo: &'c MemoPdgConstructor<'tcx, K>, ) -> DepGraph<'tcx>

pub fn to_petgraph_with_extra_global_location<'c>( &self, memo: &'c MemoPdgConstructor<'tcx, K>, extra_global_location: GlobalLocation, ) -> DepGraph<'tcx>

impl<'tcx, K> PartialGraph<'tcx, K>

pub fn new( generics: GenericArgsRef<'tcx>, def_id: DefId, arg_count: usize, local_decls: &LocalDecls<'tcx>, k: K, ) -> Self

pub(crate) fn parentable_srcs<'a>( &'a self, ) -> FxHashSet<(DepNode<'tcx, bool>, Option<u8>)>

Returns the set of source places that the parent can access (write to)

pub(crate) fn parentable_dsts<'a>( &'a self, ) -> FxHashSet<(DepNode<'tcx, bool>, Option<u8>)>

Returns the set of destination places that the parent can access (read from)