flowistry_pdg_construction
Module local_analysis

Structs

InstructionState 🔒

LocalAnalysis 🔒

Enums

CallHandling 🔒

CallKind

Functions

is_split 🔒

This used to be implemented as place_info.children(place).iter().any(|p| *p != place), but this is more efficient and should cause fewer spurious errors, since it only explores the type shallowly.

other_as_arg 🔒