Struct flowistry_pdg_construction::local_analysis::LocalAnalysis

pub(crate) struct LocalAnalysis<'tcx, 'a> {
    pub(crate) memo: &'a MemoPdgConstructor<'tcx>,
    pub(crate) root: Instance<'tcx>,
    body_with_facts: &'tcx CachedBody<'tcx>,
    pub(crate) mono_body: Body<'tcx>,
    pub(crate) def_id: DefId,
    pub(crate) place_info: PlaceInfo<'tcx>,
    control_dependencies: ControlDependencies<BasicBlock>,
    pub(crate) body_assignments: BodyAssignments,
    start_loc: FxHashSet<RichLocation>,
}

Fields
memo: &'a MemoPdgConstructor<'tcx>root: Instance<'tcx>body_with_facts: &'tcx CachedBody<'tcx>mono_body: Body<'tcx>def_id: DefIdplace_info: PlaceInfo<'tcx>control_dependencies: ControlDependencies<BasicBlock>body_assignments: BodyAssignmentsstart_loc: FxHashSet<RichLocation>
Implementations

impl<'tcx, 'a> LocalAnalysis<'tcx, 'a>

pub(crate) fn can_approximate_async_functions( &self, def_id: DefId ) -> Option<fn(_: &LocalAnalysis<'tcx, 'a>, _: &mut dyn Visitor<'tcx>, _: &[Operand<'tcx>], _: Place<'tcx>, _: Location)>

Special case behavior for calls to functions used in desugaring async functions.

Ensures that functions like Pin::new_unchecked are not modularly-approximated.

fn approximate_into_future( &self, vis: &mut dyn Visitor<'tcx>, args: &[Operand<'tcx>], destination: Place<'tcx>, location: Location )

fn approximate_new_unchecked( &self, vis: &mut dyn Visitor<'tcx>, args: &[Operand<'tcx>], destination: Place<'tcx>, location: Location )

impl<'tcx, 'mir> LocalAnalysis<'tcx, 'mir>

pub(crate) fn try_poll_call_kind<'a>( &'a self, def_id: DefId, resolved_fn: Instance<'tcx>, original_args: &'a [Operand<'tcx>] ) -> AsyncDeterminationResult<CallKind<'tcx>>

Checks whether the function call, described by the unresolved def_id and the resolved instance resolved_fn is a call to <T as Future>::poll where T is the type of an async fn or async {} created generator.

Resolves the original arguments that constituted the generator.

fn find_async_args<'a>( &'a self, args: &'a [Operand<'tcx>] ) -> Result<AsyncFnPollEnv<'tcx>, String>

Given the arguments to a Future::poll call, walk back through the body to find the original future being polled, and get the arguments to the future.

impl<'tcx, 'a> LocalAnalysis<'tcx, 'a>

pub(crate) fn new( memo: &'a MemoPdgConstructor<'tcx>, root: Instance<'tcx> ) -> LocalAnalysis<'tcx, 'a>

Creates [GraphConstructor] for a function resolved as fn_resolution in a given calling_context.

Returns None, if we were unable to load the body.

fn make_dep_node( &self, place: Place<'tcx>, location: impl Into<RichLocation> ) -> DepNode<'tcx>

pub(crate) fn find_control_inputs( &self, location: Location ) -> Vec<(DepNode<'tcx>, DepEdge)>

Returns all pairs of (src, edge)`` such that the given locationis control-dependent onedgewith inputsrc`.

fn call_change_callback(&self) -> Option<&dyn CallChangeCallback<'tcx>>

pub(crate) fn async_info(&self) -> &AsyncInfo

pub(crate) fn make_call_string( &self, location: impl Into<RichLocation> ) -> CallString

pub(crate) fn aliases( &'a self, place: Place<'tcx> ) -> impl Iterator<Item = Place<'tcx>> + 'a

Returns the aliases of place. See PlaceInfo::aliases for details.

pub(crate) fn tcx(&self) -> TyCtxt<'tcx>

pub(crate) fn find_data_inputs( &self, state: &InstructionState<'tcx>, input: Place<'tcx> ) -> Vec<DepNode<'tcx>>

Returns all nodes src such that src is:

Part of the value of input

The most-recently modified location for src

pub(crate) fn find_outputs( &self, mutated: Place<'tcx>, location: Location ) -> Vec<(Place<'tcx>, DepNode<'tcx>)>

fn apply_mutation( &self, state: &mut InstructionState<'tcx>, location: Location, mutated: Place<'tcx> )

Updates the last-mutated location for dst to the given location.

pub(crate) fn operand_to_def_id( &self, func: &Operand<'tcx> ) -> Option<(DefId, GenericArgsRef<'tcx>)>

Resolve a function Operand to a specific DefId and generic arguments if possible.

fn fmt_fn(&self, def_id: DefId) -> String

pub(crate) fn determine_call_handling<'b>( &'b self, location: Location, func: Cow<'_, Operand<'tcx>>, args: Cow<'b, [Operand<'tcx>]>, span: Span ) -> Option<CallHandling<'tcx, 'b>>

fn handle_call( &self, state: &mut InstructionState<'tcx>, location: Location, func: &Operand<'tcx>, args: &[Operand<'tcx>], destination: Place<'tcx>, span: Span ) -> bool

Attempt to inline a call to a function.

The return indicates whether we were successfully able to perform the inlining.

fn modular_mutation_visitor<'b: 'a>( &'b self, state: &'a mut InstructionState<'tcx> ) -> ModularMutationVisitor<'b, 'tcx, impl FnMut(Location, Mutation<'tcx>) + 'b>

pub(crate) fn generic_args(&self) -> GenericArgsRef<'tcx>

pub(crate) fn construct_partial(&'a self) -> PartialGraph<'tcx>

fn classify_call_kind<'b>( &'b self, def_id: DefId, resolved_fn: Instance<'tcx>, original_args: &'b [Operand<'tcx>], span: Span ) -> CallKind<'tcx>

Determine the type of call-site.

The error case is if we tried to resolve this as async and failed. We know it is async but we couldn’t determine the information needed to analyze the function, therefore we will have to approximate it.

fn try_indirect_call_kind(&self, def_id: DefId) -> Option<CallKind<'tcx>>

fn terminator_visitor<'b: 'a>( &'b self, state: &'b mut InstructionState<'tcx>, time: Time ) -> ModularMutationVisitor<'b, 'tcx, impl FnMut(Location, Mutation<'tcx>) + 'b>

impl<'tcx, 'a> LocalAnalysis<'tcx, 'a>

fn handle_terminator( &self, terminator: &Terminator<'tcx>, state: &mut InstructionState<'tcx>, location: Location, time: Time )