Crate flowistry_pdg_construction

Compute program dependence graphs (PDG) for a function call graph.

Re-exports

pub use callback::CallChangeCallback;

pub use callback::CallChangeCallbackFn;

pub use callback::CallChanges;

pub use callback::CallInfo;

pub use callback::InlineMissReason;

pub use callback::SkipCall;

Modules

approximation 🔒

async_support 🔒

body_cache

callback

Callbacks to influence graph construction and their supporting types.

calling_convention

construct 🔒

encoder

Readers and writers for the intermediate artifacts we store per crate.

graph

The representation of the PDG.

local_analysis 🔒

mutation 🔒

Identifies the mutated places in a MIR instruction via modular approximation based on types.

utils

Structs

MemoPdgConstructor

A memoizing constructor of PDGs.

Functions

compute_pdg

Computes a global program dependence graph (PDG) starting from the root function specified by def_id.

determine_async

Try to interpret this function as an async function.

is_async_trait_fn

Does this function have a structure as created by the #[async_trait] macro

match_async_trait_assign