flowistry_pdg_construction::local_analysis
Struct LocalAnalysis

pub(crate) struct LocalAnalysis<'tcx, 'a, K> {
    pub(crate) memo: &'a MemoPdgConstructor<'tcx, K>,
    pub(crate) root: Instance<'tcx>,
    body_with_facts: &'tcx CachedBody<'tcx>,
    pub(crate) mono_body: Body<'tcx>,
    pub(crate) def_id: DefId,
    pub(crate) place_info: PlaceInfo<'tcx>,
    control_dependencies: ControlDependencies<BasicBlock>,
    pub(crate) body_assignments: BodyAssignments,
    start_loc: FxHashSet<RichLocation>,
    pub(crate) param_env: TypingEnv<'tcx>,
    k: K,
}

Fields
memo: &'a MemoPdgConstructor<'tcx, K>root: Instance<'tcx>body_with_facts: &'tcx CachedBody<'tcx>mono_body: Body<'tcx>def_id: DefIdplace_info: PlaceInfo<'tcx>control_dependencies: ControlDependencies<BasicBlock>body_assignments: BodyAssignmentsstart_loc: FxHashSet<RichLocation>param_env: TypingEnv<'tcx>k: K
Implementations

impl<'tcx, 'a, K> LocalAnalysis<'tcx, 'a, K>

pub(crate) fn can_approximate_async_functions( &self, def_id: DefId, span: Span, ) -> Option<fn(_: &LocalAnalysis<'tcx, 'a, K>, _: &mut dyn Visitor<'tcx>, _: ArgSlice<'a, 'tcx>, _: Place<'tcx>, _: Location)>

Special case behavior for calls to functions used in desugaring await desugarings.

Ensures that functions like Pin::new_unchecked are not modularly-approximated.

fn approximate_into_future( &self, vis: &mut dyn Visitor<'tcx>, args: ArgSlice<'_, 'tcx>, destination: Place<'tcx>, location: Location, )

fn approximate_new_unchecked( &self, vis: &mut dyn Visitor<'tcx>, args: ArgSlice<'_, 'tcx>, destination: Place<'tcx>, location: Location, )

impl<'tcx, K> LocalAnalysis<'tcx, '_, K>

pub fn find_async_args<'a>( &'a self, resolved_fn: Instance<'tcx>, args: &'a [Spanned<Operand<'tcx>>], span: Span, ) -> Result<AsyncFnPollEnv<'tcx>, String>

Given the arguments to a Future::poll call, walk back through the body to find the original future being polled, and get the arguments to the future.

fn find_async_args_precise<'a>( &'a self, args: ArgSlice<'a, 'tcx>, ) -> Result<AsyncFnPollEnv<'tcx>, String>

impl<'tcx, 'a, K> LocalAnalysis<'tcx, 'a, K>

pub(crate) fn generic_args(&self) -> GenericArgsRef<'tcx>

fn dump_mir(&self)

pub(crate) fn find_control_inputs( &self, location: Location, ) -> Vec<(DepNode<'tcx, OneHopLocation>, DepEdge<OneHopLocation>)>

Returns all pairs of (src, edge)`` such that the given locationis control-dependent onedgewith inputsrc`.

fn call_change_callback(&self) -> &dyn CallChangeCallback<'tcx, K>

pub(crate) fn async_info(&self) -> &AsyncInfo

fn make_dep_node( &self, place: Place<'tcx>, location: impl Into<RichLocation>, ) -> DepNode<'tcx, OneHopLocation>

pub(crate) fn aliases( &'a self, place: Place<'tcx>, ) -> impl Iterator<Item = Place<'tcx>> + 'a

Returns the aliases of place. See PlaceInfo::aliases for details.

pub fn normalize_place(&self, place: &Place<'tcx>) -> Place<'tcx>

pub(crate) fn tcx(&self) -> TyCtxt<'tcx>

pub(crate) fn dcx(&self) -> DiagCtxtHandle<'tcx>

pub(crate) fn find_data_inputs( &self, state: &InstructionState<'tcx>, input: Place<'tcx>, ) -> Vec<DepNode<'tcx, OneHopLocation>>

Returns all nodes src such that src is:

Part of the value of input

The most-recently modified location for src

pub(crate) fn find_outputs( &self, mutated: Place<'tcx>, location: Location, ) -> Vec<(Place<'tcx>, DepNode<'tcx, OneHopLocation>)>

fn apply_mutation( &self, state: &mut InstructionState<'tcx>, location: Location, mutated: Place<'tcx>, )

Updates the last-mutated location for dst to the given location.

pub(crate) fn operand_to_def_id( &self, func: &Operand<'tcx>, ) -> TyAsFnResult<'tcx>

Resolve a function Operand to a specific DefId and generic arguments if possible.

fn fmt_fn(&self, def_id: DefId) -> String

impl<'tcx, 'a, K: Hash + Eq + Clone> LocalAnalysis<'tcx, 'a, K>

pub(crate) fn new( memo: &'a MemoPdgConstructor<'tcx, K>, root: Instance<'tcx>, k: K, ) -> Option<LocalAnalysis<'tcx, 'a, K>>

Creates [GraphConstructor] for a function resolved as fn_resolution in a given calling_context.

Returns None, if this body should not be analyzed.

pub(crate) fn determine_call_handling<'b>( &'b self, location: Location, func: Cow<'_, Operand<'tcx>>, args: Cow<'b, [Spanned<Operand<'tcx>>]>, span: Span, ) -> Option<CallHandling<'tcx, 'b, K>>

fn handle_call( &self, state: &mut InstructionState<'tcx>, location: Location, func: &Operand<'tcx>, args: &[Spanned<Operand<'tcx>>], destination: Place<'tcx>, span: Span, ) -> bool

Attempt to inline a call to a function.

The return indicates whether we were successfully able to perform the inlining.

fn modular_mutation_visitor<'b: 'a>( &'b self, state: &'a mut InstructionState<'tcx>, ) -> ModularMutationVisitor<'b, 'tcx, impl FnMut(Location, Mutation<'tcx>) + 'b>

pub(crate) fn construct_partial(&'a self) -> PartialGraph<'tcx, K>

fn classify_call_kind<'b>( &'b self, def_id: DefId, resolved_fn: Instance<'tcx>, original_args: &'b [Spanned<Operand<'tcx>>], span: Span, ) -> CallKind<'tcx>

Determine the type of call-site.

The error case is if we tried to resolve this as async and failed. We know it is async but we couldn’t determine the information needed to analyze the function, therefore we will have to approximate it.

In case of async, checks whether the function call, described by the unresolved def_id and the resolved instance resolved_fn is a call to <T as Future>::poll where T is the type of an async fn or async {} created generator.

Resolves the original arguments that constituted the generator.

fn try_indirect_call_kind(&self, def_id: DefId) -> Option<CallKind<'tcx>>

fn terminator_visitor<'b: 'a>( &'b self, state: &'b mut InstructionState<'tcx>, time: Time, ) -> ModularMutationVisitor<'b, 'tcx, impl FnMut(Location, Mutation<'tcx>) + 'b>

fn handle_terminator( &self, terminator: &Terminator<'tcx>, state: &mut InstructionState<'tcx>, location: Location, time: Time, )
where K: Clone + Eq + Hash,